+++ yarn-project/sequencer-client/src/client/sequencer-client.ts
@@ -1,6 +1,8 @@
 import { type L1ToL2MessageSource, type L2BlockSource, type WorldStateSynchronizer } from '@aztec/circuit-types';
 import { type ContractDataSource } from '@aztec/circuits.js';
+import { isAnvilTestChain } from '@aztec/ethereum';
 import { type EthAddress } from '@aztec/foundation/eth-address';
+import { type DateProvider } from '@aztec/foundation/timer';
 import { type P2P } from '@aztec/p2p';
 import { LightweightBlockBuilderFactory } from '@aztec/prover-client/block-builder';
 import { PublicProcessorFactory } from '@aztec/simulator';
@@ -17,7 +19,7 @@ import { TxValidatorFactory } from '../tx_validator/tx_validator_factory.js';
  * Encapsulates the full sequencer and publisher.
  */
 export class SequencerClient {
-  constructor(private sequencer: Sequencer) {}
+  constructor(protected sequencer: Sequencer) {}
 
   /**
    * Initializes and starts a new instance.
@@ -43,6 +45,7 @@ export class SequencerClient {
       l1ToL2MessageSource: L1ToL2MessageSource;
       telemetry: TelemetryClient;
       publisher?: L1Publisher;
+      dateProvider: DateProvider;
     },
   ) {
     const {
@@ -57,13 +60,30 @@ export class SequencerClient {
     const publisher = deps.publisher ?? new L1Publisher(config, telemetryClient);
     const globalsBuilder = new GlobalVariableBuilder(config);
 
-    const publicProcessorFactory = new PublicProcessorFactory(contractDataSource, telemetryClient);
+    const publicProcessorFactory = new PublicProcessorFactory(contractDataSource, deps.dateProvider, telemetryClient);
 
     const rollup = publisher.getRollupContract();
     const [l1GenesisTime, slotDuration] = await Promise.all([
       rollup.read.GENESIS_TIME(),
       rollup.read.SLOT_DURATION(),
     ] as const);
+
+    const ethereumSlotDuration = config.ethereumSlotDuration;
+
+    // When running in anvil, assume we can post a tx up until the very last second of an L1 slot.
+    // Otherwise, assume we must have broadcasted the tx before the slot started (we use a default
+    // maxL1TxInclusionTimeIntoSlot of zero) to get the tx into that L1 slot.
+    // In theory, the L1 slot has an initial 4s phase where the block is propagated, so we could
+    // make it with a propagation time into slot equal to 4s. However, we prefer being conservative.
+    // See https://www.blocknative.com/blog/anatomy-of-a-slot#7 for more info.
+    const maxL1TxInclusionTimeIntoSlot =
+      config.maxL1TxInclusionTimeIntoSlot ?? isAnvilTestChain(config.l1ChainId) ? ethereumSlotDuration : 0;
+
+    const l1Constants = {
+      l1GenesisTime,
+      slotDuration: Number(slotDuration),
+      ethereumSlotDuration,
+    };
 
     const sequencer = new Sequencer(
       publisher,
@@ -76,10 +96,10 @@ export class SequencerClient {
       l1ToL2MessageSource,
       publicProcessorFactory,
       new TxValidatorFactory(worldStateSynchronizer.getCommitted(), contractDataSource, !!config.enforceFees),
-      Number(l1GenesisTime),
-      Number(slotDuration),
+      l1Constants,
+      deps.dateProvider,
       telemetryClient,
-      config,
+      { ...config, maxL1TxInclusionTimeIntoSlot },
     );
     await validatorClient?.start();
     await sequencer.start();